02-Etl_Etl_Proyecto/data/requirements.txt
requests==2.31.0
pandas==2.2.2
python-dotenv==1.0.0
openpyxl==3.1.2
psycopg2-binary
sqlalchemy
streamlit
matplotlib==3.8.0
scikit-learn
jupyter
